+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/annotation_key_consts/annotation_key_consts_test.go
@@ -12,7 +12,7 @@ var labelKeyStrsToEnsure = map[string]string{
 }
 
 var labelKeysToEnsure = map[*kubernetes_annotation_key.KubernetesAnnotationKey]string{
-	PortSpecsAnnotationKey: "kurtosistech.com/ports",
+	PortSpecsKubernetesAnnotationKey: "kurtosistech.com/ports",
 }
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! IMPORTANT !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!