+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/api_container_object_attributes_provider.go
@@ -89,7 +89,7 @@ func (provider *kubernetesApiContainerObjectAttributesProviderImpl) ForApiContai
 
 	// Store Kurtosis port_spec info in annotation
 	annotations := map[*kubernetes_annotation_key.KubernetesAnnotationKey]*kubernetes_annotation_value.KubernetesAnnotationValue{
-		annotation_key_consts.PortSpecsAnnotationKey: serializedPortsSpec,
+		annotation_key_consts.PortSpecsKubernetesAnnotationKey: serializedPortsSpec,
 	}
 
 	objectAttributes, err := newKubernetesObjectAttributesImpl(apiContainerObjectName, labels, annotations)